requirements.txt
flask
streamlit
psycopg2-binary
requests
beautifulsoup4
pandas
sqlalchemy
